+++ instrumentation/google.golang.org/grpc/otelgrpc/benchmark_test.go
@@ -78,14 +78,6 @@ func BenchmarkUnaryServerInterceptor(b *testing.B) {
 	})
 }
 
-func BenchmarkStreamServerInterceptor(b *testing.B) {
-	benchmark(b, nil, []grpc.ServerOption{
-		grpc.StreamInterceptor(otelgrpc.StreamServerInterceptor(
-			otelgrpc.WithTracerProvider(tracerProvider),
-		)),
-	})
-}
-
 func BenchmarkUnaryClientInterceptor(b *testing.B) {
 	benchmark(b, []grpc.DialOption{
 		grpc.WithUnaryInterceptor(otelgrpc.UnaryClientInterceptor(