+++ instrumentation/google.golang.org/grpc/otelgrpc/test/interceptor_test.go
@@ -17,8 +17,6 @@ import (
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/sdk/metric"
-	"go.opentelemetry.io/otel/sdk/metric/metricdata"
-	"go.opentelemetry.io/otel/sdk/metric/metricdata/metricdatatest"
 	"go.opentelemetry.io/otel/sdk/trace"
 	"go.opentelemetry.io/otel/sdk/trace/tracetest"
 	semconv "go.opentelemetry.io/otel/semconv/v1.17.0"
@@ -885,9 +883,6 @@ func TestUnaryServerInterceptor(t *testing.T) {
 			span, ok := getSpanFromRecorder(sr, methodName)
 			require.True(t, ok, "missing span %s", methodName)
 			assertServerSpan(t, check.wantSpanCode, check.wantSpanStatusDescription, check.grpcCode, span)
-
-			// validate metric
-			assertServerMetrics(t, mr, serviceName, name, check.grpcCode)
 		})
 	}
 }
@@ -977,136 +972,6 @@ func (m *mockServerStream) SendMsg(_ interface{}) error {
 
 func (m *mockServerStream) RecvMsg(_ interface{}) error {
 	return nil
-}
-
-// TestStreamServerInterceptor tests the server interceptor for streaming RPCs.
-func TestStreamServerInterceptor(t *testing.T) {
-	for _, check := range serverChecks {
-		name := check.grpcCode.String()
-		t.Run(name, func(t *testing.T) {
-			sr := tracetest.NewSpanRecorder()
-			tp := trace.NewTracerProvider(trace.WithSpanProcessor(sr))
-
-			//nolint:staticcheck // Interceptors are deprecated and will be removed in the next release.
-			usi := otelgrpc.StreamServerInterceptor(
-				otelgrpc.WithTracerProvider(tp),
-			)
-
-			// call the stream interceptor
-			grpcErr := status.Error(check.grpcCode, check.grpcCode.String())
-			handler := func(_ interface{}, _ grpc.ServerStream) error {
-				return grpcErr
-			}
-			err := usi(&grpc_testing.SimpleRequest{}, &mockServerStream{}, &grpc.StreamServerInfo{FullMethod: name}, handler)
-			assert.Equal(t, grpcErr, err)
-
-			// validate span
-			span, ok := getSpanFromRecorder(sr, name)
-			require.True(t, ok, "missing span %s", name)
-			assertServerSpan(t, check.wantSpanCode, check.wantSpanStatusDescription, check.grpcCode, span)
-		})
-	}
-}
-
-func TestStreamServerInterceptorEvents(t *testing.T) {
-	testCases := []struct {
-		Name   string
-		Events []otelgrpc.Event
-	}{
-		{Name: "With events", Events: []otelgrpc.Event{otelgrpc.ReceivedEvents, otelgrpc.SentEvents}},
-		{Name: "With only sent events", Events: []otelgrpc.Event{otelgrpc.SentEvents}},
-		{Name: "With only received events", Events: []otelgrpc.Event{otelgrpc.ReceivedEvents}},
-		{Name: "No events", Events: []otelgrpc.Event{}},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.Name, func(t *testing.T) {
-			sr := tracetest.NewSpanRecorder()
-			tp := trace.NewTracerProvider(trace.WithSpanProcessor(sr))
-			opts := []otelgrpc.Option{
-				otelgrpc.WithTracerProvider(tp),
-			}
-			if len(testCase.Events) > 0 {
-				opts = append(opts, otelgrpc.WithMessageEvents(testCase.Events...))
-			}
-			//nolint:staticcheck // Interceptors are deprecated and will be removed in the next release.
-			usi := otelgrpc.StreamServerInterceptor(opts...)
-			stream := &mockServerStream{}
-
-			grpcCode := grpc_codes.OK
-			name := grpcCode.String()
-			// call the stream interceptor
-			grpcErr := status.Error(grpcCode, name)
-			handler := func(_ interface{}, handlerStream grpc.ServerStream) error {
-				var msg grpc_testing.SimpleRequest
-				err := handlerStream.RecvMsg(&msg)
-				require.NoError(t, err)
-				err = handlerStream.SendMsg(&msg)
-				require.NoError(t, err)
-				return grpcErr
-			}
-
-			err := usi(&grpc_testing.SimpleRequest{}, stream, &grpc.StreamServerInfo{FullMethod: name}, handler)
-			require.Equal(t, grpcErr, err)
-
-			// validate span
-			span, ok := getSpanFromRecorder(sr, name)
-			require.True(t, ok, "missing span %s", name)
-
-			if len(testCase.Events) == 0 {
-				assert.Empty(t, span.Events())
-			} else {
-				var eventsAttr []map[attribute.Key]attribute.Value
-				for _, event := range testCase.Events {
-					switch event {
-					case otelgrpc.SentEvents:
-						eventsAttr = append(eventsAttr, map[attribute.Key]attribute.Value{
-							otelgrpc.RPCMessageTypeKey: attribute.StringValue("SENT"),
-							otelgrpc.RPCMessageIDKey:   attribute.IntValue(1),
-						})
-					case otelgrpc.ReceivedEvents:
-						eventsAttr = append(eventsAttr, map[attribute.Key]attribute.Value{
-							otelgrpc.RPCMessageTypeKey: attribute.StringValue("RECEIVED"),
-							otelgrpc.RPCMessageIDKey:   attribute.IntValue(1),
-						})
-					}
-				}
-				assert.Len(t, span.Events(), len(eventsAttr))
-				assert.Equal(t, eventsAttr, eventAttrMap(span.Events()))
-			}
-		})
-	}
-}
-
-func assertServerMetrics(t *testing.T, reader metric.Reader, serviceName, name string, code grpc_codes.Code) {
-	want := metricdata.ScopeMetrics{
-		Scope: wantInstrumentationScope,
-		Metrics: []metricdata.Metrics{
-			{
-				Name:        "rpc.server.duration",
-				Description: "Measures the duration of inbound RPC.",
-				Unit:        "ms",
-				Data: metricdata.Histogram[float64]{
-					Temporality: metricdata.CumulativeTemporality,
-					DataPoints: []metricdata.HistogramDataPoint[float64]{
-						{
-							Attributes: attribute.NewSet(
-								semconv.RPCMethod(name),
-								semconv.RPCService(serviceName),
-								otelgrpc.RPCSystemGRPC,
-								otelgrpc.GRPCStatusCodeKey.Int64(int64(code)),
-							),
-						},
-					},
-				},
-			},
-		},
-	}
-	rm := metricdata.ResourceMetrics{}
-	err := reader.Collect(context.Background(), &rm)
-	assert.NoError(t, err)
-	require.Len(t, rm.ScopeMetrics, 1)
-	metricdatatest.AssertEqual(t, want, rm.ScopeMetrics[0], metricdatatest.IgnoreTimestamp(), metricdatatest.IgnoreValue())
 }
 
 func BenchmarkStreamClientInterceptor(b *testing.B) {